requirements.txt
flask
flask-pymongo
flask-cors
python-dotenv
pyjwt
bcrypt
google-generativeai
pydantic[email]
gunicorn
requests
